requirements.txt
Django==1.6.4
PIL==1.1.7
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.16
boto==2.9.9
#distribute==0.6.34
django-appconf==0.6
django-grappelli==2.5.3
django-imagekit==3.0.4
-e git+https://github.com/ninapavlich/django-inline-orderable@d1cfe9767af78d06f6edf749e7fed692178891e4#egg=django_inline_orderable-dev
django-storages==1.1.8
pilkit==1.1.9
psycopg2==2.5.1
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
